Moved functionality out of content copy service
